requirements.txt
pyserial
python-can[serial, pcan]
pynput